Allow ETag based caching of static web pages
Send the md5 hash of static files as the file etag, for all static
files.
Send a 304 Not Modified response for requests containing a
"If-None-Match" header with the correct etag.
The server now sends a Cache-Control header with the value "public,
no-cache" for static files and "no-store" for dynamic pages.

Other changes:
 * Fix embedded files always needing to be rebuilt
 * Merge Accept openmetrics and AcceptEncoding gzip check into
csvHeaderContains
 * Make uzlib_gzip_wrapper a bit more resilient against missing input